Update tut1.rst
change <table>_user to authenticated_user; grant all on schema.table to grant all on schema.

I'd like to rewrite this tutorial with a more robust example, that can include user auth where a public and private schema coexist within a database, in place of the toy example where access is granted to a specific table.  I'm find with using pgadmin to make a create tables file in tutorial 0.