Update tut1.rst
change "todo_user" that is being granted table access to "authenticated_user" that is being granted schema access.

I'd like to write a more robust tutorial that creates public and private schemas, and allows for creation of an auth/transaction system.  I'd use a pg_dump file for creating database config, and allow readers to use pgAdmin as part of tutorial.